requirements.txt
wheel~=0.36.2
aiohttp~=3.7.4.post0
aiohttp-swagger~=1.0.15
alembic~=1.6.2
async-timeout~=3.0.1
attrs~=21.2.0
certifi~=2020.12.5
chardet~=4.0.0
docopt~=0.6.2
flake8~=3.9.2
greenlet~=1.1.0
idna~=2.10
Jinja2~=2.11.3
Mako~=1.1.4
MarkupSafe~=1.1.1
mccabe~=0.6.1
multidict~=5.1.0
pipreqs~=0.4.10
psycopg2-binary~=2.8.6
pycodestyle~=2.7.0
pyflakes~=2.3.1
python-dateutil~=2.8.1
python-dotenv~=0.17.1
python-editor~=1.0.4
PyYAML~=5.4.1
requests~=2.25.1
six~=1.16.0
SQLAlchemy~=1.4.15
typing-extensions~=3.10.0.0
urllib3~=1.26.4
yarg~=0.1.9
yarl~=1.6.3
